+++ app/src/js/middleware/option-handler.js
@@ -30,6 +30,10 @@ module.exports = function (app, defaults) {
     if (req.query.theme && req.query.theme.length > 0) {
       res.opts.theme = req.query.theme;
       logger('info', `Setting Theme to ${res.opts.theme}`);
+
+      if (res.opts.theme === 'uplift') {
+        res.opts.isUpliftTheme = true;
+      }
     } else {
       res.opts.theme = 'light';
     }