+++ test/components/datagrid/datagrid.e2e-spec.js
@@ -101,7 +101,7 @@ describe('Datagrid Editable Tests', () => {
   beforeEach(async () => {
     await utils.setPage('/components/datagrid/example-editable?layout=nofrills');
 
-    const datagridEl = await element(by.id('datagrid'));
+    const datagridEl = await element(by.css('#datagrid .datagrid-body tbody tr:nth-child(1) td:nth-child(1)'));
     await browser.driver
       .wait(protractor.ExpectedConditions.presenceOf(datagridEl), config.waitsFor);
   });
@@ -123,9 +123,9 @@ describe('Datagrid Editable Tests', () => {
   it('Should render row statuses across page', async () => {
     await element(await by.id('toggle-row-status')).click();
     await element(await by.css('.pager-next a')).click();
-    await browser.driver.sleep(300);
+    await browser.driver.sleep(350);
     await element(await by.css('.pager-prev a')).click();
-    await browser.driver.sleep(300);
+    await browser.driver.sleep(350);
 
     expect(await element.all(by.css('#datagrid .rowstatus-row-error')).count()).toEqual(1);
     expect(await element.all(by.css('#datagrid .rowstatus-row-alert')).count()).toEqual(1);
@@ -137,9 +137,9 @@ describe('Datagrid Editable Tests', () => {
   it('Should not show indicator on showNewRowIndicator false', async () => {
     await element(await by.id('toggle-row-status')).click();
     await element.all(await by.css('.toolbar .btn-actions')).get(0).click();
-    await browser.driver.sleep(300);
+    await browser.driver.sleep(350);
     await element(await by.cssContainingText('li a', 'Add')).click();
-    await browser.driver.sleep(300);
+    await browser.driver.sleep(350);
 
     expect(await element.all(by.css('#datagrid .rowstatus-row-new')).count()).toEqual(0);
     expect(await element.all(by.css('#datagrid .rowstatus-row-error')).count()).toEqual(1);
@@ -248,13 +248,13 @@ describe('Datagrid filter tests', () => {
     await element(by.id('example-filter-datagrid-1-header-filter-1')).clear();
     await element(by.id('example-filter-datagrid-1-header-filter-1')).sendKeys('2241202');
     await element(by.id('example-filter-datagrid-1-header-filter-1')).sendKeys(protractor.Key.ENTER);
-    await browser.driver.sleep(300);
+    await browser.driver.sleep(350);
 
     expect(await element.all(by.css('.datagrid-body:nth-child(2) .datagrid-row')).count()).toEqual(1);
 
     await element(by.id('example-filter-datagrid-1-header-filter-1')).clear();
     await element(by.id('example-filter-datagrid-1-header-filter-1')).sendKeys(protractor.Key.ENTER);
-    await browser.driver.sleep(300);
+    await browser.driver.sleep(350);
 
     expect(await element.all(by.css('.datagrid-body:nth-child(2) .datagrid-row')).count()).toEqual(9);
     expect(await element(by.css('#datagrid .datagrid-body:nth-child(2) tbody tr:nth-child(1)')).getAttribute('class')).toMatch('is-selected');
@@ -266,7 +266,7 @@ describe('Datagrid frozen column tests', () => {
   beforeEach(async () => {
     await utils.setPage('/components/datagrid/example-frozen-columns');
 
-    const datagridEl = await element(by.css('#datagrid'));
+    const datagridEl = await element(by.css('#datagrid .datagrid-body tr:first-child'));
     await browser.driver
       .wait(protractor.ExpectedConditions.presenceOf(datagridEl), config.waitsFor);
   });
@@ -331,7 +331,7 @@ describe('Datagrid grouping with paging tests', () => {
   beforeEach(async () => {
     await utils.setPage('/components/datagrid/example-grouping-paging');
 
-    const datagridEl = await element(by.id('datagrid'));
+    const datagridEl = await element(by.css('#datagrid .datagrid-body tbody tr:nth-child(5)'));
     await browser.driver
       .wait(protractor.ExpectedConditions.presenceOf(datagridEl), config.waitsFor);
   });
@@ -394,7 +394,7 @@ describe('Datagrid index tests', () => {
   beforeEach(async () => {
     await utils.setPage('/components/datagrid/example-index?layout=nofrills');
 
-    const datagridEl = await element(by.id('datagrid'));
+    const datagridEl = await element(by.css('#datagrid tbody tr:nth-child(1)'));
     await browser.driver
       .wait(protractor.ExpectedConditions.presenceOf(datagridEl), config.waitsFor);
     await element(by.css('body')).sendKeys(protractor.Key.TAB);
@@ -505,7 +505,7 @@ describe('Datagrid mixed selection tests', () => {
   beforeEach(async () => {
     await utils.setPage('/components/datagrid/example-mixed-selection');
 
-    const datagridEl = await element(by.id('datagrid-header'));
+    const datagridEl = await element(by.css('#datagrid-header .datagrid-body:nth-child(1) tbody tr:nth-child(1) td:nth-child(2)'));
     await browser.driver
       .wait(protractor.ExpectedConditions.presenceOf(datagridEl), config.waitsFor);
   });
@@ -595,9 +595,10 @@ describe('Datagrid paging tests', () => {
   beforeEach(async () => {
     await utils.setPage('/components/datagrid/example-paging');
 
-    const datagridEl = await element(by.css('.pager-toolbar'));
+    const datagridEl = await element(by.css('#datagrid tr:nth-child(10)'));
     await browser.driver
       .wait(protractor.ExpectedConditions.presenceOf(datagridEl), config.waitsFor);
+    await browser.driver.sleep(config.sleep);
   });
 
   it('Should not have errors', async () => {
@@ -609,7 +610,7 @@ describe('Datagrid paging tests', () => {
     expect(await element(by.css('tbody tr:nth-child(10) td:nth-child(2) span')).getText()).toEqual('9');
 
     await element(by.css('.pager-last a')).click();
-    await browser.driver.sleep(300);
+    await browser.driver.sleep(config.sleep);
 
     expect(await element(by.css('tbody tr:nth-child(1) td:nth-child(2) span')).getText()).toEqual('990');
     expect(await element(by.css('tbody tr:nth-child(10) td:nth-child(2) span')).getText()).toEqual('999');
@@ -620,9 +621,9 @@ describe('Datagrid paging tests', () => {
     expect(await element(by.css('tbody tr:nth-child(10) td:nth-child(2) span')).getText()).toEqual('9');
 
     await element(by.css('.pager-last a')).click();
-    await browser.driver.sleep(300);
+    await browser.driver.sleep(config.sleep);
     await element(by.css('.pager-first a')).click();
-    await browser.driver.sleep(300);
+    await browser.driver.sleep(config.sleep);
 
     expect(await element(by.css('tbody tr:nth-child(1) td:nth-child(2) span')).getText()).toEqual('0');
     expect(await element(by.css('tbody tr:nth-child(10) td:nth-child(2) span')).getText()).toEqual('9');
@@ -633,13 +634,13 @@ describe('Datagrid paging tests', () => {
     expect(await element(by.css('tbody tr:nth-child(10) td:nth-child(2) span')).getText()).toEqual('9');
 
     await element(await by.css('.pager-next a')).click();
-    await browser.driver.sleep(300);
+    await browser.driver.sleep(config.sleep);
 
     expect(await element(by.css('tbody tr:nth-child(1) td:nth-child(2) span')).getText()).toEqual('10');
     expect(await element(by.css('tbody tr:nth-child(10) td:nth-child(2) span')).getText()).toEqual('19');
 
     await element(await by.css('.pager-prev a')).click();
-    await browser.driver.sleep(300);
+    await browser.driver.sleep(config.sleep);
 
     expect(await element(by.css('tbody tr:nth-child(1) td:nth-child(2) span')).getText()).toEqual('0');
     expect(await element(by.css('tbody tr:nth-child(10) td:nth-child(2) span')).getText()).toEqual('9');
@@ -652,31 +653,33 @@ describe('Datagrid paging tests', () => {
     await element(by.css('.pager-count input')).sendKeys(protractor.Key.BACK_SPACE);
     await element(by.css('.pager-count input')).sendKeys('5');
     await element(by.css('.pager-count input')).sendKeys(protractor.Key.ENTER);
-    await browser.driver.sleep(300);
+    await browser.driver.sleep(config.sleep);
 
     expect(await element(by.css('tbody tr:nth-child(1) td:nth-child(2) span')).getText()).toEqual('40');
     expect(await element(by.css('tbody tr:nth-child(10) td:nth-child(2) span')).getText()).toEqual('49');
   });
 
-  it('Should not move on a page thats more than the max', async () => { //eslint-disable-line
-    expect(await element(by.css('tbody tr:nth-child(1) td:nth-child(2) span')).getText()).toEqual('0');
-    expect(await element(by.css('tbody tr:nth-child(10) td:nth-child(2) span')).getText()).toEqual('9');
+  if (!utils.isCI()) {
+    it('Should not move on a page that is more than the max', async () => {
+      expect(await element(by.css('tbody tr:nth-child(1) td:nth-child(2) span')).getText()).toEqual('0');
+      expect(await element(by.css('tbody tr:nth-child(10) td:nth-child(2) span')).getText()).toEqual('9');
 
-    await element(by.css('.pager-count input')).clear();
-    await element(by.css('.pager-count input')).sendKeys('101');
-    await element(by.css('.pager-count input')).sendKeys(protractor.Key.ENTER);
-    await browser.driver.sleep(500);
+      await element(by.css('.pager-count input')).clear();
+      await element(by.css('.pager-count input')).sendKeys('101');
+      await element(by.css('.pager-count input')).sendKeys(protractor.Key.ENTER);
+      await browser.driver.sleep(config.sleep);
 
-    expect(await element(by.css('tbody tr:nth-child(1) td:nth-child(2) span')).getText()).toEqual('0');
-    expect(await element(by.css('tbody tr:nth-child(10) td:nth-child(2) span')).getText()).toEqual('9');
-  });
+      expect(await element(by.css('tbody tr:nth-child(1) td:nth-child(2) span')).getText()).toEqual('0');
+      expect(await element(by.css('tbody tr:nth-child(10) td:nth-child(2) span')).getText()).toEqual('9');
+    });
+  }
 });
 
 describe('Datagrid page size selector tests', () => {
   beforeEach(async () => {
     await utils.setPage('/components/datagrid/test-paging-page-size-selector');
 
-    const datagridEl = await element(by.id('datagrid'));
+    const datagridEl = await element(by.css('#datagrid .datagrid-body tbody tr:nth-child(1) td:nth-child(1)'));
     await browser.driver
       .wait(protractor.ExpectedConditions.presenceOf(datagridEl), config.waitsFor);
   });
@@ -735,16 +738,16 @@ describe('Datagrid single select tests', () => {
 
     // Sort
     await element(by.css('#datagrid .datagrid-header th:nth-child(4)')).click();
-    await browser.driver.sleep(300);
+    await browser.driver.sleep(350);
     await element(by.css('#datagrid .datagrid-header th:nth-child(4)')).click();
 
     expect(await element(by.css('#datagrid .datagrid-row.is-selected td:nth-child(1) span')).getText()).toEqual('2142201');
 
-    await browser.driver.sleep(300);
+    await browser.driver.sleep(350);
 
     await element(by.css('#datagrid .datagrid-body tbody tr:nth-child(1) td:nth-child(1)')).click();
 
-    await browser.driver.sleep(300);
+    await browser.driver.sleep(350);
 
     expect(await element(by.css('#datagrid .datagrid-body tbody tr:nth-child(1)')).getAttribute('class')).toMatch('is-selected');
     expect(await element(by.css('#datagrid .datagrid-body tbody tr:nth-child(2)')).getAttribute('class')).not.toMatch('is-selected');
@@ -765,7 +768,7 @@ describe('Datagrid Client Side Filter and Sort Tests', () => {
     await element(by.css('#datagrid thead th:nth-child(2) .datagrid-header-text')).click();
 
     expect(await element(by.css('#datagrid thead th:nth-child(2)')).getText()).toEqual('Product Id');
-    await browser.driver.sleep(300);
+    await browser.driver.sleep(350);
 
     expect(await element(by.css('#datagrid thead th:nth-child(2)')).getAttribute('class')).toContain('is-sorted-asc');
   });
@@ -781,7 +784,7 @@ describe('Datagrid Client Side Filter and Sort Tests', () => {
 
   it('Should retain both filter and sort criteria', async () => {
     await element(by.css('#datagrid thead th:nth-child(2) .datagrid-header-text')).click();
-    await browser.driver.sleep(300);
+    await browser.driver.sleep(350);
     await element(by.css('#datagrid thead th:nth-child(2) input')).sendKeys('22');
     await browser.driver.sleep(500);
 
@@ -1079,7 +1082,7 @@ describe('Datagrid disableRowDeactivation setting tests', () => {
   beforeEach(async () => {
     await utils.setPage('/components/datagrid/test-mixed-selection-disable-row-dectivation');
 
-    const datagridEl = await element(by.id('datagrid-header'));
+    const datagridEl = await element(by.css('#datagrid-header .datagrid-body tbody tr:nth-child(1)'));
     await browser.driver
       .wait(protractor.ExpectedConditions.presenceOf(datagridEl), config.waitsFor);
   });
@@ -1170,7 +1173,7 @@ describe('Datagrid paging multiselect across pages', () => {
   beforeEach(async () => {
     await utils.setPage('/components/datagrid/test-paging-multiselect-select-across-page');
 
-    const datagridEl = await element(by.id('datagrid'));
+    const datagridEl = await element(by.css('#datagrid .datagrid-body tbody tr:nth-child(2)'));
     await browser.driver
       .wait(protractor.ExpectedConditions.presenceOf(datagridEl), config.waitsFor);
   });
@@ -1204,7 +1207,7 @@ describe('Datagrid paging multiselect tests', () => {
   beforeEach(async () => {
     await utils.setPage('/components/datagrid/test-paging-multiselect');
 
-    const datagridEl = await element(by.id('datagrid'));
+    const datagridEl = await element(by.css('#datagrid .datagrid-body tbody tr:nth-child(1)'));
     await browser.driver
       .wait(protractor.ExpectedConditions.presenceOf(datagridEl), config.waitsFor);
   });
@@ -1285,14 +1288,12 @@ describe('Datagrid paging clientside single select tests', () => {
 
     expect(await element.all(by.css('.datagrid-row.is-selected')).count()).toEqual(1);
 
-    await element(await by.css('.pager-next')).click();
-
+    await element(await by.css('.pager-next a')).click();
     await browser.driver.sleep(config.sleep);
 
     expect(await element.all(by.css('.datagrid-row.is-selected')).count()).toEqual(0);
 
-    await element(by.css('.pager-prev')).click();
-
+    await element(by.css('.pager-prev a')).click();
     await browser.driver.sleep(config.sleep);
 
     expect(await element.all(by.css('.datagrid-row.is-selected')).count()).toEqual(1);
@@ -1318,14 +1319,12 @@ describe('Datagrid paging indeterminate multiple select tests', () => {
 
     expect(await element.all(await by.css('.datagrid-row.is-selected')).count()).toEqual(2);
 
-    await element(await by.css('.pager-next')).click();
-
+    await element(await by.css('.pager-next a')).click();
     await browser.driver.sleep(config.sleep);
 
     expect(await element.all(await by.css('.datagrid-row.is-selected')).count()).toEqual(0);
 
-    await element(await by.css('.pager-prev')).click();
-
+    await element(await by.css('.pager-prev a')).click();
     await browser.driver.sleep(config.sleep);
 
     expect(await element.all(by.css('.datagrid-row.is-selected')).count()).toEqual(0);
@@ -1351,17 +1350,13 @@ describe('Datagrid paging indeterminate single select tests', () => {
 
     expect(await element.all(by.css('.datagrid-row.is-selected')).count()).toEqual(1);
 
-    await element(await by.css('.pager-next')).click();
-
-    await browser.driver
-      .wait(protractor.ExpectedConditions.elementToBeClickable(await element(by.css('.pager-prev'))), config.waitsFor);
+    await element(await by.css('.pager-next a')).click();
+    await browser.driver.sleep(config.sleep);
 
     expect(await element.all(by.css('.datagrid-row.is-selected')).count()).toEqual(0);
 
-    await element(by.css('.pager-prev')).click();
-
-    await browser.driver
-      .wait(protractor.ExpectedConditions.elementToBeClickable(await element(by.css('.pager-next'))), config.waitsFor);
+    await element(by.css('.pager-prev a')).click();
+    await browser.driver.sleep(config.sleep);
 
     expect(await element.all(by.css('.datagrid-row.is-selected')).count()).toEqual(0);
   });
@@ -1656,7 +1651,7 @@ describe('Datagrid tree do not select siblings tests', () => {
   beforeEach(async () => {
     await utils.setPage('/components/datagrid/test-tree-select-siblings');
 
-    const datagridEl = await element(by.id('datagrid'));
+    const datagridEl = await element(by.css('#datagrid .datagrid-body tbody tr:nth-child(1) td:nth-child(1)'));
     await browser.driver
       .wait(protractor.ExpectedConditions.presenceOf(datagridEl), config.waitsFor);
   });
@@ -1738,6 +1733,17 @@ describe('Datagrid tooltip tests', () => {
       .wait(protractor.ExpectedConditions.visibilityOf(await element(by.css('.grid-tooltip'))), config.waitsFor);
     tooltip = await element(by.css('.grid-tooltip'));
 
+    expect(await tooltip.getAttribute('class')).not.toContain('is-hidden');
+  });
+
+  it('Should show tooltip on header text cut off with ellipsis', async () => {
+    await browser.actions().mouseMove(element(by.css('.datagrid-header thead th[data-column-id="orderDate"] .datagrid-column-wrapper'))).perform();
+    await browser.driver.sleep(config.sleep);
+    await browser.driver
+      .wait(protractor.ExpectedConditions.presenceOf(await element(by.css('.grid-tooltip'))), config.waitsFor);
+    const tooltip = await element(by.css('.grid-tooltip'));
+
+    expect(await element(by.css('.datagrid-header thead th[data-column-id="orderDate"]')).getAttribute('class')).toContain('text-ellipsis');
     expect(await tooltip.getAttribute('class')).not.toContain('is-hidden');
   });
 });